+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 1.3.0
+- All color handling is now handled by Prism library.
+  - Now supports web colors for easier color definitions.
+  - Now supports animated colors for both borders and backgrounds.
+  - Now supports automatic border colors, which is the new default.  This means any border images can have automatically determined border and background colors that match the image.
+- The color section in the configuration file is now in a more compact format.  Configuration files in the old format will be automatically converted.
+- Both start and end background colors can now be specified separately.
+
 ### 1.2.5
 - Fixed a bug that caused excessive configuration file backups.